Rename to portfolio_valuation, move functions to pools_ namespace
