+++ src/rpc/pools.rs
@@ -201,4 +201,30 @@ where
 			.map_err(|e| runtime_error("Unable to query tranche currency.", e))?
 			.ok_or_else(|| invalid_params_error("Pool or tranche not found."))
 	}
+
+	fn portfolio_valuation(&self, pool_id: PoolId, at: Option<Block::Hash>) -> RpcResult<Balance> {
+		let api = self.client.runtime_api();
+		let at = if let Some(hash) = at {
+			BlockId::hash(hash)
+		} else {
+			BlockId::hash(self.client.info().best_hash)
+		};
+
+		api.pool_valuation(&at, pool_id)
+			.map_err(|e| runtime_error("Unable to query pool valuation.", e))?
+			.ok_or_else(|| invalid_params_error("Pool not found."))
+	}
+
+	fn max_borrow_amount(&self, pool_id: PoolId, at: Option<Block::Hash>) -> RpcResult<Balance> {
+		let api = self.client.runtime_api();
+		let at = if let Some(hash) = at {
+			BlockId::hash(hash)
+		} else {
+			BlockId::hash(self.client.info().best_hash)
+		};
+
+		api.max_borrow_amount(&at, pool_id)
+			.map_err(|e| runtime_error("Unable to query pool valuation.", e))?
+			.ok_or_else(|| invalid_params_error("Pool not found."))
+	}
 }
